refactor: extract log controllers
