+++ src/features/attack-log/components/isActivationKey.ts
@@ -1,0 +1,2 @@
+export const isActivationKey = (key: string) =>
+  key === 'Enter' || key === ' ' || key === 'Spacebar';
