+++ src/features/attack-log/types.ts
@@ -1,0 +1,8 @@
+export type AttackLogActionPayload = {
+  id: string;
+  label: string;
+  damage: number;
+  category: string;
+  soulCost: number | null;
+};
+
